23.12.06 ranking, mypage
1. ranking 오류 수정
2. mypage date 오류 수정
3. mypage 레벨 관련 조언 추가
--- 완성 ---
